+++ src/main/java/delfos/dataset/basic/rating/IteratorRatingsDataset.java
@@ -32,7 +32,7 @@ import java.util.Map;
 public class IteratorRatingsDataset<RatingType extends Rating> implements Iterator<RatingType> {
 
     private RatingType _next;
-    private final List<Integer> _users;
+    private final List<Long> _users;
     private final List<RatingType> _ratings;
     private final RatingsDataset<RatingType> _ratingsDataset;
     private final Object exMut;
@@ -82,16 +82,16 @@ public class IteratorRatingsDataset<RatingType extends Rating> implements Iterat
                 _next = null;
             } else {
                 //Hay mas usuarios, cargar sus ratings.
-                int idUser = _users.remove(0);
+                long idUser = _users.remove(0);
 
-                Map<Integer, RatingType> userRatingsRated = _ratingsDataset.getUserRatingsRated(idUser);
+                Map<Long, RatingType> userRatingsRated = _ratingsDataset.getUserRatingsRated(idUser);
 
                 while (userRatingsRated.isEmpty()) {
                     idUser = _users.remove(0);
                     userRatingsRated = _ratingsDataset.getUserRatingsRated(idUser);
                 }
 
-                for (Map.Entry<Integer, RatingType> entry : userRatingsRated.entrySet()) {
+                for (Map.Entry<Long, RatingType> entry : userRatingsRated.entrySet()) {
                     _ratings.add(entry.getValue());
                 }
